requirements.txt
# local package
-e .

# external requirements
docutils==0.15.2
Sphinx==4.2.0
sphinx-rtd-theme==0.5.2
numpydoc==1.1.0
pytest==6.2.5
black==21.8b0
pre-commit==2.15.0
PyYAML==5.4.1
awscli==1.20.40
pandas==1.3.3
Scrapy==2.5.0
streamlit==0.88.0
